split QIRGeneration and clean up
